src/manager/requirements.txt
amqp==1.4.9
anyjson==0.3.3
atomicwrites==1.4.0
attrs==21.4.0
billiard==3.3.0.23
bkoauth==0.0.25
blueapps==3.6.1rc0
celery==3.1.25
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.2
croniter==1.1.0
cryptography==3.1.1
ddtrace==0.14.1
django==2.2.16
django-celery==3.2.1
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.13.1
drf-ujson2==1.6.1
drf-yasg==1.20.0
faker==8.8.2
fakeredis==1.7.0
gunicorn==19.6.0
idna==2.8
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
jinja2==2.11.3
kombu==3.0.37
mako==1.0.6
markupsafe==1.1.1
more-itertools==8.12.0
msgpack-python==0.5.6
mysqlclient==1.4.4
packaging==21.3
pluggy==0.13.1
py==1.11.0
pycparser==2.21
pycryptodome==3.14.1
pyjwt==1.6.1
pymongo==3.11.0
pyparsing==3.0.7
pytest==5.4.3
pytest-cov==2.12.1
pytest-django==3.9.0
pytest-sugar==0.9.3
python-dateutil==2.8.2
python-json-logger==0.1.11
pytz==2021.3
raven==6.5.0
redis==2.10.5
requests==2.22.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.2
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
typing-extensions==4.1.1
ujson==4.3.0
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.5
werkzeug==0.14.1
whitenoise==3.3.1
wrapt==1.13.3
zipp==3.6.0
